@@ -11,20 +11,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^15.2.10",
-    "@angular/animations": "^15.2.10",
-    "@angular/cdk": "^14.2.9",
-    "@angular/cli": "^15.2.10",
-    "@angular/common": "^15.2.10",
-    "@angular/compiler": "^15.2.10",
-    "@angular/core": "^15.2.10",
-    "@angular/forms": "^15.2.10",
-    "@angular/localize": "^15.2.10",
-    "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "^15.2.10",
-    "@angular/platform-browser-dynamic": "^15.2.10",
-    "@angular/router": "^15.2.10",
-    "@ctrl/ngx-codemirror": "4.1.1",
+    "@angular-devkit/schematics": "^16.2.10",
+    "@angular/animations": "^16.2.12",
+    "@angular/cdk": "16.2.12",
+    "@angular/cli": "^16.2.10",
+    "@angular/common": "^16.2.12",
+    "@angular/compiler": "^16.2.12",
+    "@angular/core": "^16.2.12",
+    "@angular/forms": "^16.2.12",
+    "@angular/localize": "^16.2.12",
+    "@angular/material": "16.2.12",
+    "@angular/platform-browser": "^16.2.12",
+    "@angular/platform-browser-dynamic": "^16.2.12",
+    "@angular/router": "^16.2.12",
+    "@ctrl/ngx-codemirror": "^7.0.0",
     "@sentry/angular": "^6.3.6",
     "@sentry/tracing": "^6.3.6",
     "@types/mousetrap": "^1.6.5",
@@ -37,12 +37,12 @@
     "ngx-toastr": "16.2.0",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.10",
-    "@angular/compiler-cli": "^15.2.10",
-    "@angular/language-service": "^15.2.10",
+    "@angular-devkit/build-angular": "^16.2.10",
+    "@angular/compiler-cli": "^16.2.12",
+    "@angular/language-service": "^16.2.12",
     "@babel/preset-env": "^7.11.5",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
