@@ -1,37 +1,38 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import {MatDividerModule} from '@angular/material/divider'
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatTableModule } from '@angular/material/table';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatButtonModule } from '@angular/material/button';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatInputModule } from '@angular/material/input';
-import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatRadioModule } from '@angular/material/radio';
-import { MatSelectModule } from '@angular/material/select';
-import { MatSliderModule } from '@angular/material/slider';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatSidenavModule } from '@angular/material/sidenav';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatListModule } from '@angular/material/list';
-import { MatGridListModule } from '@angular/material/grid-list';
-import { MatCardModule } from '@angular/material/card';
-import { MatStepperModule } from '@angular/material/stepper';
-import { MatTabsModule } from '@angular/material/tabs';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { MatButtonToggleModule } from '@angular/material/button-toggle';
-import { MatChipsModule } from '@angular/material/chips';
-import { MatIconModule } from '@angular/material/icon';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatSnackBarModule  } from '@angular/material/snack-bar';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatNativeDateModule} from '@angular/material/core';
-import { from } from 'rxjs';
-import { HttpClientModule} from '@angular/common/http';
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { MatDividerModule } from "@angular/material/divider";
+
+import { MatPseudoCheckboxModule } from "@angular/material/core";
+import { MatDatepickerModule } from "@angular/material/datepicker";
+import { MatSidenavModule } from "@angular/material/sidenav";
+import { MatToolbarModule } from "@angular/material/toolbar";
+import { MatGridListModule } from "@angular/material/grid-list";
+import { MatStepperModule } from "@angular/material/stepper";
+import { MatExpansionModule } from "@angular/material/expansion";
+import { MatButtonToggleModule } from "@angular/material/button-toggle";
+import { MatIconModule } from "@angular/material/icon";
+import { MatFormFieldModule } from "@angular/material/form-field";
+import { MatDialogModule } from "@angular/material/dialog";
+import { MatInputModule } from "@angular/material/input";
+import { MatSelectModule } from "@angular/material/select";
+import { MatButtonModule } from "@angular/material/button";
+import { MatSnackBarModule } from "@angular/material/snack-bar";
+import { MatMenuModule } from "@angular/material/menu";
+import { MatCheckboxModule } from "@angular/material/checkbox";
+import { MatTableModule } from "@angular/material/table";
+import { MatAutocompleteModule } from "@angular/material/autocomplete";
+import { MatSlideToggleModule } from "@angular/material/slide-toggle";
+import { MatSliderModule } from "@angular/material/slider";
+import { MatRadioModule } from "@angular/material/radio";
+import { MatListModule } from "@angular/material/list";
+import { MatCardModule } from "@angular/material/card";
+import { MatTabsModule } from "@angular/material/tabs";
+import { MatProgressSpinnerModule } from "@angular/material/progress-spinner";
+import { MatProgressBarModule } from "@angular/material/progress-bar";
+import { MatTooltipModule } from "@angular/material/tooltip";
+import { MatChipsModule } from "@angular/material/chips";
+import { HttpClientModule } from "@angular/common/http";
+import { MatNativeDateModule } from "@angular/material/core";
 
 @NgModule({
   declarations: [],
@@ -44,6 +45,7 @@ import { HttpClientModule} from '@angular/common/http';
     MatButtonModule,
     MatDialogModule,
     MatCheckboxModule,
+    MatPseudoCheckboxModule,
     MatInputModule,
     MatSnackBarModule,
     MatAutocompleteModule,
@@ -67,20 +69,18 @@ import { HttpClientModule} from '@angular/common/http';
     MatProgressSpinnerModule,
     MatProgressBarModule,
     MatTooltipModule,
-    HttpClientModule, 
-    MatNativeDateModule
+    HttpClientModule,
+    MatNativeDateModule,
   ],
   exports: [
     MatDividerModule,
     MatCheckboxModule,
+    MatPseudoCheckboxModule,
     MatTableModule,
     MatFormFieldModule,
     MatButtonModule,
     MatDialogModule,
-    MatCheckboxModule,
-    MatButtonModule,
     MatInputModule,
-
     MatAutocompleteModule,
     MatDatepickerModule,
     MatRadioModule,
@@ -103,8 +103,7 @@ import { HttpClientModule} from '@angular/common/http';
     MatProgressBarModule,
     MatTooltipModule,
     HttpClientModule,
-    MatNativeDateModule
-  ]
-
+    MatNativeDateModule,
+  ],
 })
-export class MaterialModule { }
+export class MaterialModule {}
